crtg_voice_agent/requirements.txt
fastapi
uvicorn
twilio
google-generativeai
elevenlabs
deepgram-sdk
python-dotenv
websockets
requests
pydantic
